feat: :sparkles: displays the mangas on the home page
