--- mangaReader/src/props/componentProps.ts
@@ -1,13 +1,0 @@
-export type ButtonProp = {
-  name: string;
-};
-
-export type MangaCardProp = {
-  id: string;
-};
-
-export type singleMangaProp = {
-  mangaId: string | undefined;
-  title: string | undefined;
-  coverFileName: string | undefined;
-};
