+++ mangaReader/src/props/componentTypes.ts
@@ -1,0 +1,21 @@
+export type ButtonProp = {
+  name: string;
+};
+
+export type MangaCardProp = {
+  id: string;
+};
+
+export type singleMangaProp = {
+  mangaId: string;
+  title: string;
+  coverID: string;
+};
+
+export type HeadProps = {
+  tags: string[];
+};
+
+export type NavProps = {
+  name: string;
+};
